+++ lib/models/agora_rtc_event_handlers.dart
@@ -140,13 +140,12 @@ class AgoraRtcEventHandlers {
   /// Occurs when the first local video frame is displayed on the local video view.
   /// The SDK triggers this callback when the first local video frame is displayed on the local video view.
   ///
-  /// * [source] The capture type of the custom video source. See VideoSourceType .
-  /// * [connection] The connection information. See RtcConnection .
+  /// * [source] The type of the video source. See VideoSourceType .
   /// * [width] The width (px) of the first local video frame.
   /// * [height] The height (px) of the first local video frame.
-  /// * [elapsed] Time elapsed (ms) from the local user calling joinChannel [2/2] until the SDK triggers this callback. If you call startPreview before calling joinChannel [2/2], then this parameter is the time elapsed from calling the startPreview method until the SDK triggers this callback.
+  /// * [elapsed] Time elapsed (ms) from the local user calling joinChannel until the SDK triggers this callback. If you call startPreview before calling joinChannel, then this parameter is the time elapsed from calling the startPreview method until the SDK triggers this callback.
   final void Function(
-          RtcConnection connection, int width, int height, int elapsed)?
+          VideoSourceType source, int width, int height, int elapsed)?
       onFirstLocalVideoFrame;
 
   /// Occurs when the first video frame is published.
@@ -273,15 +272,6 @@ class AgoraRtcEventHandlers {
   final void Function(RtcConnection connection, int remoteUid, bool enabled)?
       onUserEnableLocalVideo;
 
-  /// Occurs when a method is executed by the SDK.
-  ///
-  ///
-  /// * [err] The error code returned by the SDK when the method call fails. If the SDK returns 0, then the method call is successful.
-  /// * [api] The method executed by the SDK.
-  /// * [result] The result of the method call.
-  final void Function(ErrorCodeType err, String api, String result)?
-      onApiCallExecuted;
-
   /// Reports the statistics of the local audio stream.
   /// The SDK triggers this callback once every two seconds.
   ///
@@ -340,8 +330,12 @@ class AgoraRtcEventHandlers {
   /// * [vecRectangle] The information of the detected human face. See Rectangle .
   /// * [vecDistance] The distance between the human face and the device screen (cm).
   /// * [numFaces] The number of faces detected. If the value is 0, it means that no human face is detected.
-  final void Function(int imageWidth, int imageHeight, Rectangle vecRectangle,
-      int vecDistance, int numFaces)? onFacePositionChanged;
+  final void Function(
+      int imageWidth,
+      int imageHeight,
+      List<Rectangle> vecRectangle,
+      List<int> vecDistance,
+      int numFaces)? onFacePositionChanged;
 
   /// Occurs when the video stops playing.
   /// Deprecated:Use localVideoStreamStateStopped(0) in the onLocalVideoStateChanged callback instead.The application can use this callback to change the configuration of the view (for example, displaying other pictures in the view) after the video stops playing.
@@ -777,7 +771,6 @@ class AgoraRtcEventHandlers {
     this.onUserEnableVideo,
     this.onUserStateChanged,
     this.onUserEnableLocalVideo,
-    this.onApiCallExecuted,
     this.onLocalAudioStats,
     this.onRemoteAudioStats,
     this.onLocalVideoStats,